Refactor payment features and remove unused event class
Deleted `PaymentRejectedEvent` as it is no longer in use. Renamed folder and namespace `Querys` to `Queries` for consistency. Introduced `ICreateOrderCommandSubmittedQueueClient` interface and moved related messaging queue client to the `Payments` folder for improved organization and separation of concerns.
